Move the database logic into its own module Move the database logic out of the inventory stale updater command
