+++ client_id_to_account_id_map.json
@@ -1,5 +1,5 @@
 {
-    "client-0": "010101",
+    "8a823918-f801-474c-9f79-7cc7e9d79a7f": "010101",
     "client-1": "010102"
 }
