+++ cmd/cloud-connector/inventory_stale_timestamp_updater.go
@@ -2,11 +2,10 @@ package main
 
 import (
 	"context"
-	"database/sql"
-	"encoding/json"
 	"time"
 
 	"github.com/RedHatInsights/cloud-connector/internal/config"
+	"github.com/RedHatInsights/cloud-connector/internal/connection_repository"
 	"github.com/RedHatInsights/cloud-connector/internal/controller"
 	"github.com/RedHatInsights/cloud-connector/internal/domain"
 	"github.com/RedHatInsights/cloud-connector/internal/platform/db"
@@ -17,73 +16,6 @@ import (
 )
 
 const serviceName = "Cloud-Connector Inventory Stale Timestamp Updater"
-
-type connectionProcessor func(context.Context, domain.RhcClient) error
-
-func processStaleConnections(ctx context.Context, databaseConn *sql.DB, sqlTimeout time.Duration, staleTimeCutoff time.Time, chunkSize int, processConnection connectionProcessor) error {
-
-	logger.Log.Debug("Host's should be updated if their stale_timestamp is before ", staleTimeCutoff)
-
-	queryCtx, cancel := context.WithTimeout(ctx, sqlTimeout)
-	defer cancel()
-
-	statement, err := databaseConn.Prepare(
-		`SELECT account, client_id, canonical_facts, tags FROM connections
-           WHERE canonical_facts != '{}' AND
-             dispatchers ? 'rhc-worker-playbook' AND
-             stale_timestamp < $1
-             order by stale_timestamp asc
-             limit $2`)
-	if err != nil {
-		logger.LogFatalError("SQL Prepare failed", err)
-		return nil
-	}
-	defer statement.Close()
-
-	rows, err := statement.QueryContext(queryCtx, staleTimeCutoff, chunkSize)
-	if err != nil {
-		logger.LogFatalError("SQL query failed", err)
-		return nil
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var account domain.AccountID
-		var clientID domain.ClientID
-		var canonicalFactsString string
-		var tagsString string
-
-		if err := rows.Scan(&account, &clientID, &canonicalFactsString, &tagsString); err != nil {
-			logger.LogError("SQL scan failed.  Skipping row.", err)
-			continue
-		}
-
-		var canonicalFacts interface{}
-		err = json.Unmarshal([]byte(canonicalFactsString), &canonicalFacts)
-		if err != nil {
-			logger.LogErrorWithAccountAndClientId("Unable to parse canonical facts.  Skipping connection.", err, account, clientID)
-			continue
-		}
-
-		var tags domain.Tags
-		err = json.Unmarshal([]byte(tagsString), &tags)
-		if err != nil {
-			logger.LogErrorWithAccountAndClientId("Unable to parse tags.  Skipping connection.", err, account, clientID)
-			continue
-		}
-
-		rhcClient := domain.RhcClient{
-			Account:        account,
-			ClientID:       clientID,
-			CanonicalFacts: canonicalFacts,
-			Tags:           tags,
-		}
-
-		processConnection(ctx, rhcClient)
-	}
-
-	return nil
-}
 
 func calculateStaleCutoffTime(staleTimeOffset time.Duration) time.Time {
 	windowToUpdateBeforeGoingStale := 1 * time.Hour
@@ -129,7 +61,9 @@ func startInventoryStaleTimestampUpdater() {
 	tooOldIfBeforeThisTime := calculateStaleCutoffTime(cfg.InventoryStaleTimestampOffset)
 	chunkSize := cfg.InventoryStaleTimestampUpdaterChunkSize
 
-	processStaleConnections(context.TODO(), databaseConn, sqlTimeout, tooOldIfBeforeThisTime, chunkSize,
+	logger.Log.Debug("Host's should be updated if their stale_timestamp is before ", tooOldIfBeforeThisTime)
+
+	connection_repository.ProcessStaleConnections(context.TODO(), databaseConn, sqlTimeout, tooOldIfBeforeThisTime, chunkSize,
 		func(ctx context.Context, rhcClient domain.RhcClient) error {
 
 			log := logger.Log.WithFields(logrus.Fields{"client_id": rhcClient.ClientID, "account": rhcClient.Account})
@@ -149,7 +83,7 @@ func startInventoryStaleTimestampUpdater() {
 				return err
 			}
 
-			updateStaleTimestampInDB(log, ctx, databaseConn, sqlTimeout, rhcClient)
+			connection_repository.UpdateStaleTimestampInDB(log, ctx, databaseConn, sqlTimeout, rhcClient)
 
 			return nil
 		})
@@ -158,32 +92,4 @@ func startInventoryStaleTimestampUpdater() {
 	if err := kafkaProducer.Close(); err != nil {
 		logger.LogFatalError("Failed to close the kafka writer", err)
 	}
-}
-
-func updateStaleTimestampInDB(log *logrus.Entry, ctx context.Context, databaseConn *sql.DB, sqlTimeout time.Duration, rhcClient domain.RhcClient) {
-
-	log.Debug("Updating stale timestamp")
-
-	ctx, cancel := context.WithTimeout(ctx, sqlTimeout)
-	defer cancel()
-
-	update := "UPDATE connections SET stale_timestamp = NOW() WHERE account=$1 AND client_id=$2"
-
-	statement, err := databaseConn.Prepare(update)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer statement.Close()
-
-	results, err := statement.ExecContext(ctx, rhcClient.Account, rhcClient.ClientID)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	rowsAffected, err := results.RowsAffected()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	log.Debug("rowsAffected:", rowsAffected)
 }
