+++ internal/connection_repository/sql_connection_locator.go
@@ -1,4 +1,4 @@
-package controller
+package connection_repository
 
 import (
 	"context"
@@ -6,6 +6,7 @@ import (
 	"encoding/json"
 
 	"github.com/RedHatInsights/cloud-connector/internal/config"
+	"github.com/RedHatInsights/cloud-connector/internal/controller"
 	"github.com/RedHatInsights/cloud-connector/internal/domain"
 	"github.com/RedHatInsights/cloud-connector/internal/platform/db"
 	"github.com/RedHatInsights/cloud-connector/internal/platform/logger"
@@ -16,7 +17,7 @@ import (
 
 type SqlConnectionLocator struct {
 	database     *sql.DB
-	proxyFactory ReceptorProxyFactory
+	proxyFactory controller.ReceptorProxyFactory
 	metrics      *sqlConnectionLookupMetrics
 }
 
@@ -47,7 +48,7 @@ func initializeSqlConnectionLookupMetrics() *sqlConnectionLookupMetrics {
 	return metrics
 }
 
-func NewSqlConnectionLocator(cfg *config.Config, proxyFactory ReceptorProxyFactory) (*SqlConnectionLocator, error) {
+func NewSqlConnectionLocator(cfg *config.Config, proxyFactory controller.ReceptorProxyFactory) (*SqlConnectionLocator, error) {
 
 	database, err := db.InitializeDatabaseConnection(cfg)
 	if err != nil {
@@ -61,8 +62,8 @@ func NewSqlConnectionLocator(cfg *config.Config, proxyFactory ReceptorProxyFacto
 	}, nil
 }
 
-func (scm *SqlConnectionLocator) GetConnection(ctx context.Context, account domain.AccountID, client_id domain.ClientID) Receptor {
-	var conn Receptor
+func (scm *SqlConnectionLocator) GetConnection(ctx context.Context, account domain.AccountID, client_id domain.ClientID) controller.Receptor {
+	var conn controller.Receptor
 	var err error
 
 	callDurationTimer := prometheus.NewTimer(scm.metrics.sqlLookupConnectionByAccountAndClientIDDuration)
@@ -101,12 +102,12 @@ func (scm *SqlConnectionLocator) GetConnection(ctx context.Context, account doma
 	return conn
 }
 
-func (scm *SqlConnectionLocator) GetConnectionsByAccount(ctx context.Context, account domain.AccountID) map[domain.ClientID]Receptor {
+func (scm *SqlConnectionLocator) GetConnectionsByAccount(ctx context.Context, account domain.AccountID) map[domain.ClientID]controller.Receptor {
 
 	callDurationTimer := prometheus.NewTimer(scm.metrics.sqlLookupConnectionsByAccountDuration)
 	defer callDurationTimer.ObserveDuration()
 
-	connectionsPerAccount := make(map[domain.ClientID]Receptor)
+	connectionsPerAccount := make(map[domain.ClientID]controller.Receptor)
 
 	statement, err := scm.database.Prepare("SELECT client_id, dispatchers FROM connections WHERE account = $1")
 	if err != nil {
@@ -148,12 +149,12 @@ func (scm *SqlConnectionLocator) GetConnectionsByAccount(ctx context.Context, ac
 	return connectionsPerAccount
 }
 
-func (scm *SqlConnectionLocator) GetAllConnections(ctx context.Context) map[domain.AccountID]map[domain.ClientID]Receptor {
+func (scm *SqlConnectionLocator) GetAllConnections(ctx context.Context) map[domain.AccountID]map[domain.ClientID]controller.Receptor {
 
 	callDurationTimer := prometheus.NewTimer(scm.metrics.sqlLookupAllConnectionsDuration)
 	defer callDurationTimer.ObserveDuration()
 
-	connectionMap := make(map[domain.AccountID]map[domain.ClientID]Receptor)
+	connectionMap := make(map[domain.AccountID]map[domain.ClientID]controller.Receptor)
 
 	statement, err := scm.database.Prepare("SELECT account, client_id, dispatchers FROM connections")
 	if err != nil {
@@ -191,7 +192,7 @@ func (scm *SqlConnectionLocator) GetAllConnections(ctx context.Context) map[doma
 		}
 
 		if _, exists := connectionMap[account]; !exists {
-			connectionMap[account] = make(map[domain.ClientID]Receptor)
+			connectionMap[account] = make(map[domain.ClientID]controller.Receptor)
 		}
 
 		connectionMap[account][clientId] = proxy
