+++ internal/connection_repository/sql_connection_registrar.go
@@ -1,4 +1,4 @@
-package controller
+package connection_repository
 
 import (
 	"context"
