+++ internal/controller/api/message_receiver.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 
 	"github.com/RedHatInsights/cloud-connector/internal/config"
+	"github.com/RedHatInsights/cloud-connector/internal/connection_repository"
 	"github.com/RedHatInsights/cloud-connector/internal/controller"
 	"github.com/RedHatInsights/cloud-connector/internal/domain"
 	"github.com/RedHatInsights/cloud-connector/internal/middlewares"
@@ -15,13 +16,13 @@ import (
 )
 
 type MessageReceiver struct {
-	connectionMgr controller.ConnectionLocator
+	connectionMgr connection_repository.ConnectionLocator
 	router        *mux.Router
 	config        *config.Config
 	urlPrefix     string
 }
 
-func NewMessageReceiver(cm controller.ConnectionLocator, r *mux.Router, urlPrefix string, cfg *config.Config) *MessageReceiver {
+func NewMessageReceiver(cm connection_repository.ConnectionLocator, r *mux.Router, urlPrefix string, cfg *config.Config) *MessageReceiver {
 	return &MessageReceiver{
 		connectionMgr: cm,
 		router:        r,
